fix: move react component to seperate file
