+++ src/pages/api/receipt/download.ts
@@ -4,6 +4,7 @@ import sharp from "sharp";
 import { readFile } from "node:fs/promises";
 import { fileURLToPath } from "node:url";
 import { generateSummaryReceipt, generateCommitsReceipt } from "../../../lib/receipt";
+import { ReceiptImage } from "../../../lib/ReceiptImage";
 
 let fontCache: ArrayBuffer | null = null;
 
@@ -13,54 +14,6 @@ async function loadFont(): Promise<ArrayBuffer> {
     const buffer = await readFile(fontPath);
     fontCache = buffer.buffer.slice(buffer.byteOffset, buffer.byteOffset + buffer.byteLength);
     return fontCache;
-}
-
-function createReceiptElement(receiptText: string, scale: number): React.ReactNode {
-    const lines = receiptText.split("\n");
-
-    return {
-        type: "div",
-        props: {
-            style: {
-                display: "flex",
-                flexDirection: "column",
-                alignItems: "center",
-                justifyContent: "center",
-                padding: 48 * scale,
-                backgroundColor: "#0d0d0d",
-                width: "100%",
-                height: "100%",
-            },
-            children: {
-                type: "div",
-                props: {
-                    style: {
-                        display: "flex",
-                        flexDirection: "column",
-                        fontFamily: "Fira Code",
-                        fontSize: 14 * scale,
-                        lineHeight: 1.4,
-                        color: "#e0e0e0",
-                        backgroundColor: "#0d0d0d",
-                        padding: 32 * scale,
-                        borderRadius: 8 * scale,
-                        border: `${scale}px solid #333`,
-                    },
-                    children: lines.map((line, i) => ({
-                        type: "div",
-                        key: String(i),
-                        props: {
-                            style: {
-                                display: "flex",
-                                whiteSpace: "pre",
-                            },
-                            children: line || " ",
-                        },
-                    })),
-                },
-            },
-        },
-    };
 }
 
 export const GET: APIRoute = async ({ url }) => {
@@ -87,7 +40,7 @@ export const GET: APIRoute = async ({ url }) => {
         const width = 700;
         const height = Math.max(400, lines.length * 20 + 160);
 
-        const svg = await satori(createReceiptElement(receiptText, scale), {
+        const svg = await satori(ReceiptImage({ receiptText, scale }), {
             width: width * scale,
             height: height * scale,
             fonts: [
@@ -104,7 +57,7 @@ export const GET: APIRoute = async ({ url }) => {
             .png()
             .toBuffer();
 
-        return new Response(pngBuffer, {
+        return new Response(new Uint8Array(pngBuffer), {
             status: 200,
             headers: {
                 "Content-Type": "image/png",
